+++ include/beast/http/read.hpp
@@ -52,17 +52,13 @@ namespace http {
 
     @param parser The parser to use.
 
-    @return The number of octets processed from the dynamic buffer.
-    The octets should be removed by calling `consume` on the dynamic
-    buffer after the read completes, regardless of any error.
-
     @throws system_error Thrown on failure.
 */
 template<
     class SyncReadStream,
     class DynamicBuffer,
     bool isRequest, class Derived>
-std::size_t
+void
 read_some(
     SyncReadStream& stream,
     DynamicBuffer& buffer,
@@ -108,16 +104,12 @@ read_some(
     @param parser The parser to use.
 
     @param ec Set to the error, if any occurred.
-
-    @return The number of octets processed from the dynamic buffer.
-    The octets should be removed by calling `consume` on the dynamic
-    buffer after the read completes, regardless of any error.
 */
 template<
     class SyncReadStream,
     class DynamicBuffer,
     bool isRequest, class Derived>
-std::size_t
+void
 read_some(
     SyncReadStream& stream,
     DynamicBuffer& buffer,
@@ -168,8 +160,7 @@ read_some(
     completes. Copies will be made of the handler as required.
     The equivalent function signature of the handler must be:
     @code void handler(
-        error_code const& error,    // result of operation
-        std::size_t bytes_used      // the number of bytes to consume
+        error_code const& error     // result of operation
     ); @endcode
     Regardless of whether the asynchronous operation completes
     immediately or not, the handler will not be invoked from within
@@ -190,7 +181,7 @@ template<
     void_or_deduced
 #else
 async_return_type<
-    ReadHandler, void(error_code, std::size_t)>
+    ReadHandler, void(error_code)>
 #endif
 async_read_some(
     AsyncReadStream& stream,